Ansible pull fails in interesting ways.
The error message:
ERROR! The inventory file 'ansible/hosts' looks like it should be an executable inventory script, but is not marked executable. Perhaps you want to correct this with `chmod +x ansible/hosts`?

is obscured by:
UnicodeDecodeError: 'ascii' codec can't decode byte 0xe2 in position 226: ordinal not in range(128)